requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
datasketch==1.6.4
elastic-transport==8.10.0
elasticsearch==8.10.1
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
packaging==23.2
parsel==1.8.1
Pillow==10.1.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
PyDispatcher==2.0.7
pyOpenSSL==23.3.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
scipy==1.10.1
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tldextract==5.0.1
torch==2.1.0
torch-vision==0.1.6.dev0
triton==2.1.0
Twisted==22.10.0
typing-extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0
w3lib==2.1.2
zope.interface==6.1
datasketch==1.6.4